RankFragment 관련 부분 이름 변경 및 구조 일부 변경
