--- app/src/main/java/com/soopeach/thethethe_android/RankFragment.kt
@@ -1,88 +1,0 @@
-package com.soopeach.thethethe_android
-
-import android.content.Context
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.soopeach.thethethe_android.databinding.FragmentRankBinding
-import com.soopeach.thethethe_android.databinding.RankRetrofitBinding
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-/**
- * A simple [Fragment] subclass.
- * Use the [RankFragment.newInstance] factory method to
- * create an instance of this fragment.
- */
-
-class RankViewHolder(val binding: RankRetrofitBinding) : RecyclerView.ViewHolder(binding.root)
-class RankAdapter(val context: Context) : RecyclerView.Adapter<RecyclerView.ViewHolder>(){
-    override fun getItemCount(): Int {
-        return 10
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder
-    =RankViewHolder(RankRetrofitBinding.inflate(LayoutInflater.from(parent.context),parent,false))
-
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
-        val binding=(holder as RankViewHolder).binding
-    }
-}
-class RankFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        val binding =FragmentRankBinding.inflate(inflater, container, false)
-
-        binding.RecyclerView.layoutManager = LinearLayoutManager(requireContext())
-        binding.RecyclerView.adapter = RankAdapter(requireContext())
-
-        binding.rankingMy
-        binding.coupleMemoMy
-        binding.coupleNameMy
-
-
-        return binding.root
-    }
-
-    companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment RankFragment.
-         */
-        // TODO: Rename and change types and number of parameters
-        @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            RankFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
-    }
-}
