+++ app/src/main/res/layout/rank_item.xml
@@ -1,58 +1,61 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="#F5F5F5"
     android:orientation="vertical"
     android:paddingTop="10dp"
-    android:paddingBottom="0dp">
-
+    android:paddingBottom="0dp"
+    tools:ignore="MissingDefaultResource">
 
 
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:background="@color/white"
         android:paddingLeft="10dp"
-        android:paddingTop="10dp"
-        android:background="@color/white">
+        android:paddingTop="10dp">
 
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="10dp"
-        android:layout_marginRight="10dp"
-        android:id="@+id/option1"
-        >
-        <TextView
-            android:id="@+id/ranking"
+        <RelativeLayout
+            android:id="@+id/option1"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="등수"/>
+            android:layout_marginLeft="10dp"
+            android:layout_marginRight="10dp">
 
-        <ImageView
-            android:layout_toRightOf="@+id/ranking"
-            android:layout_marginLeft="20dp"
-            android:layout_width="50dp"
-            android:layout_height="50dp"
-            android:src="@drawable/user_profile"
-            />
-    </RelativeLayout>
+            <TextView
+                android:id="@+id/ranking"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="등수" />
+
+            <com.soopeach.thethethe_android.custom.FlexibleCornerImageView
+                android:id="@+id/couple_img"
+                android:layout_width="50dp"
+                android:layout_height="50dp"
+                app:all_corner_radius="25dp"
+                android:layout_marginLeft="20dp"
+                android:layout_toRightOf="@+id/ranking"
+                android:src="@drawable/user_profile" />
+        </RelativeLayout>
 
         <RelativeLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginLeft="10dp"
             android:layout_marginRight="10dp"
-            android:layout_toRightOf="@+id/option1"
-            >
+            android:layout_toRightOf="@+id/option1">
+
             <TextView
                 android:id="@+id/couple_name"
                 android:layout_width="220dp"
                 android:layout_height="wrap_content"
                 android:text="사용자"
-                android:textStyle="bold"
                 android:textColor="#000000"
-                android:textSize="20sp" />
+                android:textSize="20sp"
+                android:textStyle="bold" />
 
             <TextView
                 android:id="@+id/item_type"
@@ -63,6 +66,7 @@
                 android:text="장소 종류"
                 android:textColor="#727272"
                 android:textSize="10sp" />
+
             <TextView
                 android:id="@+id/couple_memo"
                 android:layout_width="match_parent"
@@ -74,27 +78,25 @@
                 android:textSize="14sp" />
 
             <RelativeLayout
+                android:id="@+id/save_btn"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:visibility="visible"
-                android:id="@+id/save_btn">
+                android:visibility="visible">
 
                 <TextView
                     android:id="@+id/point"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="점수"
                     android:layout_marginLeft="200dp"
                     android:layout_marginTop="13sp"
-                    android:textSize="15sp"
-                    android:textColor="#537CC5"/>
+                    android:text="점수"
+                    android:textColor="#537CC5"
+                    android:textSize="15sp" />
 
             </RelativeLayout>
 
 
         </RelativeLayout>
-
-
 
 
     </RelativeLayout>
